% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_hourly_means_fast.R
\name{calculate_hourly_means_fast}
\alias{calculate_hourly_means_fast}
\title{Function to calculate hourly means quickly for an \strong{smonitor} 
database.}
\usage{
calculate_hourly_means_fast(con, process, start, end, tz = "UTC")
}
\arguments{
\item{con}{Database connection.}

\item{process}{A vector of processes.}

\item{start}{Start date to import.}

\item{end}{End date to import.}

\item{tz}{Time-zone for the dates to be parsed into. Default is \code{"UTC"}.}
}
\description{
Function to calculate hourly means quickly for an \strong{smonitor} 
database.
}
\author{
Stuart K. Grange
}

