% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_monthly_summaries.R
\name{calculate_monthly_summaries}
\alias{calculate_monthly_summaries}
\title{Function to calculate monthly means of observations in an \strong{smonitor}
database.}
\usage{
calculate_monthly_summaries(con, df, verbose = TRUE, progress = "none")
}
\arguments{
\item{con}{Database connection.}

\item{df}{Mapping data frame which contains \code{"site"} and 
\code{"process"} variables.}

\item{verbose}{Should the function print what site and processes are being 
processed?}

\item{progress}{What type of progress bar should be displayed?}
}
\description{
Function to calculate monthly means of observations in an \strong{smonitor}
database.
}
\author{
Stuart K.Grange
}
