% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_process_spans.R
\name{calculate_process_spans}
\alias{calculate_process_spans}
\title{Function to find start and end dates of a process in an \strong{smonitor}
database.}
\usage{
calculate_process_spans(con, tz = "UTC")
}
\arguments{
\item{con}{Database connection.}

\item{tz}{Time-zone to parse the dates to.}
}
\description{
To-do: Add WHERE clauses for summary and validity and also processes.
}
\author{
Stuart K.Grange
}

