% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_summaries.R
\name{calculate_summaries}
\alias{calculate_summaries}
\alias{calculate_summaries_efficently}
\title{Function to calculate time series summaries for data stored in a 
\strong{smonitor} database.}
\usage{
calculate_summaries(
  con,
  df,
  start = NA,
  end = NA,
  tz = "UTC",
  insert = TRUE,
  verbose = FALSE
)

calculate_summaries_efficently(
  con,
  df,
  start = NA,
  end = NA,
  tz = "UTC",
  insert = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{con}{Database connection.}

\item{df}{A mapping table containing \code{"process"}, \code{"source_summary"},
\code{"summary"}, \code{"summary_name"}, \code{"period"}, 
\code{"aggregation_function"}, \code{"validity_threshold"}, and 
\code{"source_name_summary"} variables.}

\item{start}{Start date for summaries.}

\item{end}{End date for summaries.}

\item{tz}{Time zone in use.}

\item{insert}{Should the data be inserted? This is a useful option to have
to allow for a dry run of an insert.}

\item{verbose}{Should the function give messages?}
}
\value{
Invisible database connection.
}
\description{
\code{calculate_summaries} will handle aggregation methods, transformation, 
and updating processes when the \strong{smonitor} tables are complete.
}
\details{
\code{calculate_summaries} will correctly aggregate wind direction.
}
\examples{
\dontrun{

# Calculate all summaries in data_look_up for 2014 and insert them into the 
# database
calculate_summaries(con, data_look_up, start = 2014, start = 2014)

}

}
\seealso{
\code{\link{import_summaries}}, \code{\link[openair]{timeAverage}}, 
\code{\link[dplyr]{summarise}}, \code{\link[threadr]{time_pad}}, 
\code{\link[threadr]{aggregate_by_date}}
}
\author{
Stuart K. Grange
}
