% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_summaries_efficently.R
\name{calculate_summaries_efficently}
\alias{calculate_summaries_efficently}
\title{Function to calculate time-series summaries for data stored in a 
\strong{smonitor} database.}
\usage{
calculate_summaries_efficently(con, df_map, start = NA, end = NA,
  tz = "UTC", insert = TRUE, verbose = FALSE, progress = "none")
}
\arguments{
\item{con}{Database connection.}

\item{df_map}{A mapping table containing \code{"process"}, \code{"summary"},
\code{"aggregation_function"}, and \code{"validity_threshold"} variables.}

\item{start}{Start date for summaries.}

\item{end}{End date for summaries.}

\item{insert}{Should the data be inserted? Default is \code{TRUE}.}

\item{verbose}{Should the function be verbose?}

\item{progress}{Type of progress bar to display.}
}
\value{
Invisible.
}
\description{
\code{calculate_summaries_efficently} will handle aggregation methods, 
transformation, and updating processes when the \strong{smonitor} tables are
complete.
}
\details{
\code{calculate_summaries_efficently} will correctly aggregate wind direction.
}
\examples{
\dontrun{

# Calculate all summaries in data_look_up for 2014 and insert them into the 
# database
calculate_summaries_efficently(con, data_look_up, start = 2014, start = 2014)

}

}
\seealso{


\code{\link{import_summaries}}, \code{\link{timeAverage}}, 
\code{\link{summarise}}, \code{\link{time_pad}}, 
\code{\link{aggregate_by_date}}
}
\author{
Stuart K. Grange
}
