% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_process.R
\name{check_process}
\alias{check_process}
\title{Function to test if processes exist in the `processes` table.}
\usage{
check_process(con, process)
}
\arguments{
\item{con}{Database connection to a smonitor database.}

\item{process}{A vector of processes to test.}
}
\value{
Integer vector, if length is 0, the processes already exist.
}
\description{
Function to test if processes exist in the `processes` table.
}
\author{
Stuart K. Grange
}
