% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_process_vector.R
\name{create_new_process_vector}
\alias{create_new_process_vector}
\title{Function to generate a vector of \strong{smonitor} process keys.}
\usage{
create_new_process_vector(con, n, minimum = NA, positive = TRUE)
}
\arguments{
\item{con}{Database connection to an \strong{smonitor} database.}

\item{n}{Number of process keys to generate.}

\item{minimum}{If \code{positive}, what is the minimum value of the process 
key? Useful for when process keys want to be grouped together.}

\item{positive}{Should the next process key be greater than the current 
process in the database? If \code{FALSE}, negative sequences are usually 
to be generated.}
}
\value{
Integer vector with the length of \code{n}.
}
\description{
Function to generate a vector of \strong{smonitor} process keys.
}
\author{
Stuart K. Grange
}
