% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_observations_table.R
\name{create_observations_table}
\alias{create_observations_table}
\title{Function to create \code{`observations`} table for a \strong{smonitor} 
database.}
\usage{
create_observations_table(con, index = TRUE)
}
\arguments{
\item{con}{Database connection.}

\item{index}{Should an index on the \code{process} variable be created? 
Default is \code{TRUE}.}
}
\description{
Function to create \code{`observations`} table for a \strong{smonitor} 
database.
}
\author{
Stuart K. Grange
}

