% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary_entries.R
\name{create_summary_entries}
\alias{create_summary_entries}
\title{Function to create entries for the \code{`summary`} table in a 
\strong{smonitor} database.}
\usage{
create_summary_entries(
  con,
  process,
  type = "standard",
  validity_threshold = 0L
)
}
\arguments{
\item{con}{Database connection.}

\item{process}{A process integer.}

\item{type}{Type of entries to create. Can be either \code{"standard"} or
\code{"extended"}.}

\item{validity_threshold}{Threshold for valid average.}
}
\description{
Function to create entries for the \code{`summary`} table in a 
\strong{smonitor} database.
}
\author{
Stuart K. Grange
}
