% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_missing_observations.R
\name{delete_missing_observations}
\alias{delete_missing_observations}
\title{Function to delete missing (\code{NA}/\code{NULL}) value from the 
\code{`observations`} table in an \strong{smonitor} database.}
\usage{
delete_missing_observations(con)
}
\arguments{
\item{con}{Database connection.}
}
\value{
Invisible.
}
\description{
Function to delete missing (\code{NA}/\code{NULL}) value from the 
\code{`observations`} table in an \strong{smonitor} database.
}
\author{
Stuart K. Grange.
}
