% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_aurn_data.R
\name{download_aurn}
\alias{download_aurn}
\title{Function to get observations from the United Kingdom's Automatic Urban and 
Rural Network (AURN) with \strong{openair's} \code{importAURN} function.}
\usage{
download_aurn(site, start = 1990, end = NA)
}
\arguments{
\item{site}{Site code, for example \code{"yk11"}.}

\item{start}{Start year.}

\item{end}{End year.}
}
\description{
Function to get observations from the United Kingdom's Automatic Urban and 
Rural Network (AURN) with \strong{openair's} \code{importAURN} function.
}
\author{
Stuart K. Grange
}
