% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_kings_college_data.R
\name{download_kings_college}
\alias{download_kings_college}
\title{Function to get observations which are served by Kings College London and 
accessed with \strong{openair}.}
\usage{
download_kings_college(site, start, end = NA)
}
\arguments{
\item{start}{Start year.}

\item{end}{End year.}

\item{Site}{Site code, for example \code{"rb1"}.}
}
\description{
Function to get observations which are served by Kings College London and 
accessed with \strong{openair}.
}
\author{
Stuart K. Grange
}

