% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_worldmet_data.R
\name{download_noaa}
\alias{download_noaa}
\title{Function to get observations from NOAA's Integrated Surface Database (ISD)
with \code{worldmet}.}
\usage{
download_noaa(site, start = 1990, end = NA)
}
\arguments{
\item{site}{Site code, for example \code{"037720-99999"}.}

\item{start}{Start year.}

\item{end}{End year.}
}
\description{
Function to get observations from NOAA's Integrated Surface Database (ISD)
with \code{worldmet}.
}
\author{
Stuart K. Grange
}
