% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump_observations.R
\name{dump_observations}
\alias{dump_observations}
\title{Function to export an \strong{smonitor}'s \code{`observations`} table as 
\code{.rds} objects.}
\usage{
dump_observations(con, n = 100, output, progress = "time")
}
\arguments{
\item{con}{Database connection.}

\item{n}{Number of processes to query and save for each iteration. Default is
\code{100}.}

\item{output}{Output directory to where \code{.rds} objects should be saved.}

\item{progress}{Type of progress bar to display. Default is \code{"time"}.}
}
\description{
Function to export an \strong{smonitor}'s \code{`observations`} table as 
\code{.rds} objects.
}
\examples{
\dontrun{
dump_observations(con, n = 5, output = "~/Desktop/")
}

}
\author{
Stuart K. Grange
}

