% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_observations_by_site.R
\name{export_observations_by_site}
\alias{export_observations_by_site}
\title{Function to export tidy tabular files from a \strong{smonitor} database which 
can be shared and easily used.}
\usage{
export_observations_by_site(con, df, start = 1960, end = NA,
  site_name = FALSE, directory_output,
  file_name_prefix = NA_character_, file_type = ".csv.gz",
  verbose = FALSE)
}
\arguments{
\item{con}{Database connection to a \strong{smonitor} database.}

\item{df}{A mapping data frame containing \code{site} and \code{process} 
variables.}

\item{start}{Start date of observations to query from database.}

\item{end}{End date of observations to query from database.}

\item{site_name}{Should the site name variable be included in the return?}

\item{directory_output}{The directory to export files to.}

\item{file_name_prefix}{Prefix to use for the exported file names. The default
is \code{"air_quality_data_site_"}.}

\item{file_type}{Type of files to export. Can be either \code{".csv.gz"} or 
\code{".rds"}.}

\item{verbose}{Should the function give messages?}
}
\value{
Invisible \code{con} and exported files.
}
\description{
\code{export_observations_by_site} will create site-based data files.
}
\author{
Stuart K. Grange
}
