% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_missing_process_keys.R
\name{find_missing_process_keys}
\alias{find_missing_process_keys}
\title{Function to get missing process keys for a \strong{smonitor} database.}
\usage{
find_missing_process_keys(con)
}
\arguments{
\item{con}{Database connection to an \strong{smonitor} database.}
}
\value{
Integer vector.
}
\description{
Function to get missing process keys for a \strong{smonitor} database.
}
\author{
Stuart K. Grange
}
