% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aurn_data.R
\name{get_aurn_data}
\alias{get_aurn_data}
\title{Function to get UK's AURN data using the same API as \strong{openair}.}
\usage{
get_aurn_data(site, year, longer = FALSE, print_query = FALSE)
}
\arguments{
\item{site}{Vector of sites.}

\item{year}{Vector of years.}

\item{longer}{Should the data frame be reshaped to be longer with the use of
\strong{tidyr}'s \code{gather} function?}

\item{print_query}{Should the function print the URLs which are being loaded?}
}
\value{
Tibble.
}
\description{
Function to get UK's AURN data using the same API as \strong{openair}.
}
\author{
Stuart K. Grange
}
