% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_process_integer.R
\name{get_latest_process_integer}
\alias{get_latest_process_integer}
\title{Function to get latest process integer from an \strong{smonitor} database.}
\usage{
get_latest_process_integer(con)
}
\arguments{
\item{con}{Database connection.}
}
\description{
Function to get latest process integer from an \strong{smonitor} database.
}
\author{
Stuart K. Grange
}
