% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_processes_sequence.R
\name{get_processes_sequence}
\alias{get_processes_sequence}
\title{Function to generate a sequence of process keys and test if they exist in
an \strong{smonitor} database.}
\usage{
get_processes_sequence(con, only_outstanding = FALSE)
}
\arguments{
\item{con}{A \strong{smonitor} database connection.}

\item{only_outstanding}{Should the return be filtered to only contain the 
processes that do not exist in the database?}
}
\value{
Tibble.
}
\description{
\code{get_processes_sequence} is useful in determining if groups of processes
have been deleted.
}
\author{
Stuart K. Grange
}
