% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_aggregations.R
\name{import_aggregations}
\alias{import_aggregations}
\title{Function to import \code{`aggregations`} table from a \strong{smonitor} 
database.}
\usage{
import_aggregations(con, extra = TRUE)
}
\arguments{
\item{con}{Database connection.}

\item{extra}{Return extra data? Default is \code{TRUE}.}
}
\description{
Function to import \code{`aggregations`} table from a \strong{smonitor} 
database.
}

