% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_any.R
\name{import_any}
\alias{import_any}
\title{Function to import observational data from a \strong{smonitor} database.}
\usage{
import_any(con, process, summary = NA, start = 1970, end = NA,
  tz = "UTC", valid_only = TRUE, date_end = TRUE, date_insert = FALSE,
  site_name = TRUE, print_query = FALSE)
}
\arguments{
\item{con}{A \strong{smonitor} database connection.}

\item{process}{A vector of processes.}

\item{summary}{A vector of summaries. If summary is \code{NA} (the default), 
then only the \code{process} will be used in the \code{WHERE} clause.}

\item{start}{What is the start date of data to be returned? Ideally, the 
date format should be \code{yyyy-mm-dd}, but the UK locale convention of 
\code{dd/mm/yyyy} will also work. Years as strings or integers work too and
will floor-rounded.}

\item{end}{What is the end date of data to be returned? Ideally, the 
date format should be \code{yyyy-mm-dd}, but the UK locale convention of 
\code{dd/mm/yyyy} will also work. Years as strings or integers work too and 
will be ceiling-rounded.}

\item{tz}{Time-zone for the dates to be parsed into. Default is \code{"UTC"}.}

\item{valid_only}{Should invalid observations be filtered out? Default is 
\code{TRUE}. Valid observations are considered to be those with the validity
variable being \code{1} or missing (\code{NULL} or \code{NA}).}

\item{date_end}{Should the return include the \code{date_end} variable? 
Default is \code{TRUE}.}

\item{date_insert}{Should the return include the \code{date_insert} variable? 
Default is \code{TRUE}.}

\item{site_name}{Should the return include the \code{site_name} variable? 
Default is \code{TRUE}.}

\item{print_query}{Should the SQL query string be printed?}
}
\value{
Data frame containing decoded observataion data with correct data 
types.
}
\description{
\code{import_any} is considered the primary data importer for 
\strong{smonitor}.
}
\author{
Stuart K. Grange
}
\seealso{
\code{\link{import_by_site}}
}

