% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_by_process.R
\name{import_by_process}
\alias{import_by_process}
\title{Function to import observational data from a \strong{smonitor} database.}
\usage{
import_by_process(con, process = NA, summary = NA, start = 1969,
  end = NA, tz = "UTC", valid_only = TRUE, date_end = TRUE,
  date_insert = FALSE, site_name = TRUE, unit = TRUE, warn = TRUE,
  print_query = FALSE)
}
\arguments{
\item{con}{A \strong{smonitor} database connection.}

\item{process}{A vector of processes.}

\item{summary}{A vector of summaries. If summary is \code{NA} (the default), 
then only the \code{process} will be used in the \code{WHERE} clause.}

\item{start}{What is the start date of data to be returned? Ideally, the 
date format should be \code{yyyy-mm-dd}, but the UK locale convention of 
\code{dd/mm/yyyy} will also work. Years as strings or integers work too and
will floor-rounded.}

\item{end}{What is the end date of data to be returned? Ideally, the 
date format should be \code{yyyy-mm-dd}, but the UK locale convention of 
\code{dd/mm/yyyy} will also work. Years as strings or integers work too and 
will be ceiling-rounded.}

\item{tz}{Time-zone for the dates to be parsed into. Default is \code{"UTC"}.}

\item{valid_only}{Should invalid observations be filtered out? Default is 
\code{TRUE}. Valid observations are considered to be those with the validity
variable being \code{1} or missing (\code{NULL} or \code{NA}).}

\item{date_end}{Should the return include the \code{date_end} variable? 
Default is \code{TRUE}.}

\item{date_insert}{Should the return include the \code{date_insert} variable? 
Default is \code{FALSE}.}

\item{site_name}{Should the return include the \code{site_name} variable? 
Default is \code{TRUE}.}

\item{unit}{Should the processes' units be included in the return?}

\item{warn}{Should the functions raise warnings?}

\item{print_query}{Should the SQL query string be printed?}
}
\value{
Data frame containing decoded observational data with correct data 
types.
}
\description{
\code{import_by_process} is considered the primary data importer for 
\strong{smonitor}.
}
\examples{
\dontrun{

# Simple usage, all summaries for a process will be returned
data_example <- import_by_process(con, process = 400)

# Many processes
data_example <- import_by_process(con, process = 1:20)

# Simple usage, a single summary will be returned
data_example <- import_by_process(con, process = 400, summary = 1)

}

}
\seealso{
\code{\link{import_by_site}} for a higher-level importing function
}
\author{
Stuart K. Grange
}
