% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_by_site.R
\name{import_by_site}
\alias{import_by_site}
\title{Function to import data from an \code{smonitor} database based on sites.}
\usage{
import_by_site(con, site, start = 1970, end = NA, period = "hour",
  valid_only = FALSE, pad = TRUE, tz = "UTC", spread = FALSE,
  europe = FALSE)
}
\arguments{
\item{con}{A \strong{smonitor} database connection.}

\item{site}{A site code such as \code{"my1"}.}

\item{start}{Start date to import.}

\item{end}{End date to import.}

\item{period}{Averaging period. Default is \code{"hour"}}

\item{valid_only}{Should only valid data be returned? Default is \code{FALSE}.}

\item{pad}{Should the time-series be padded to ensure all dates in the 
observation period are present? Default is \code{TRUE}.}

\item{tz}{Time-zone for the dates to be parsed into. Default is \code{"UTC"}.}

\item{spread}{Should the data frame take the wider format resulting from
spreading the data? Default is \code{FALSE}.}

\item{europe}{Should \code{import_processes_europe} be used rather than
\code{import_processes}?}
}
\description{
Function to import data from an \code{smonitor} database based on sites.
}
\author{
Stuart K. Grange
}

