% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_by_zone.R
\name{import_by_zone}
\alias{import_by_zone}
\title{Function to import data from an \strong{smonitor} database based on zones/
spatial polygons.}
\usage{
import_by_zone(con, zone, variable = NA, sp_other = NA, start = 1969,
  end = NA, period = "hour", pad = TRUE, spread = TRUE, tz = "UTC",
  valid_only = TRUE, print_query = FALSE)
}
\arguments{
\item{con}{Database connection.}

\item{zone}{A vector of zones to use as the boundaries for site filtering. 
Use \code{\link{import_zone_names}} to get the vectors needed.}

\item{variable}{An optional variable vector. If not used, all variables will
be selected and returned.}

\item{sp_other}{An optional spatial polygon which exists outside the database
tables to be used as a boundary rather than using \code{zone}.}

\item{start}{Start date to import.}

\item{end}{End date to import.}

\item{period}{Averaging period. Default is \code{"hour"}. \code{period} can
also take the value \code{"any"} which will return all periods, but 
\code{pad} argument will be ignored.}

\item{pad}{Should the time-series be padded to ensure all dates in the 
observation period are present? Default is \code{TRUE}.}

\item{spread}{Should the data frame take the wider format resulting from
spreading the data? Default is \code{FALSE}.}

\item{tz}{Time-zone for the dates to be parsed into. Default is \code{"UTC"}.}

\item{valid_only}{Should only valid data be returned? Default is \code{TRUE}.}

\item{print_query}{Should the SQL query string be printed?}
}
\description{
\code{import_by_zone} needs a \code{`zones`} table and has been developed 
specifically for \strong{smonitor} Europe and it therefore not generic. To-do:
make it generic.
}
\author{
Stuart K. Grange
}
\seealso{
\code{\link{import_zones}}, \code{\link{import_zone_names}}
}

