% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_processes_europe.R
\name{import_country_codes}
\alias{import_country_codes}
\title{Function to insert country codes from the European \strong{smonitor} database.}
\usage{
import_country_codes(con)
}
\arguments{
\item{con}{Database connection.}
}
\description{
Function to insert country codes from the European \strong{smonitor} database.
}
\author{
Stuart K. Grange
}

