% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_europe_sites_local.R
\name{import_europe_processes_local}
\alias{import_europe_processes_local}
\title{Function to import a local copy of the \strong{smonitor} Europe 
\code{processes} table.}
\usage{
import_europe_processes_local(file = NA, tz = "UTC")
}
\arguments{
\item{file}{Local file name.}

\item{tz}{Time zone for dates.}
}
\value{
Data frame.
}
\description{
Function to import a local copy of the \strong{smonitor} Europe 
\code{processes} table.
}
\author{
Stuart K. Grange
}
