% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_functions.R
\name{import_hourly_means}
\alias{import_hourly_means}
\title{Function to import hourly means from a database.}
\usage{
import_hourly_means(con, process)
}
\arguments{
\item{con}{Database connection.}

\item{process}{A process, an integer key.}
}
\description{
Function to import hourly means from a database.
}
\author{
Stuart K. Grange
}

