% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_invalidations.R
\name{import_invalidations}
\alias{import_invalidations}
\title{Function to import \code{`invalidations`} table from a \strong{smonitor} 
database.}
\usage{
import_invalidations(con, tz = "UTC")
}
\arguments{
\item{con}{Database connection.}

\item{tz}{What time zone should the \code{date_start} and \code{date_end}
variables be represented as?}
}
\value{
Data frame.
}
\description{
Function to import \code{`invalidations`} table from a \strong{smonitor} 
database.
}
\examples{
\dontrun{

# Import invalidations from a smonitor database
data_invalidations <- import_invalidations(con)

}

}
\author{
Stuart K. Grange
}
