% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_monthly_summaries.R
\name{import_monthly_summaries}
\alias{import_monthly_summaries}
\title{Function to import monthly summaries from an \strong{smonitor} database.}
\usage{
import_monthly_summaries(con, site = NA, variable = NA, period = NA,
  date_insert = FALSE, spread = FALSE, tz = "UTC")
}
\arguments{
\item{con}{Database connection.}

\item{site}{Sites to import.}

\item{variable}{Variables to import.}

\item{period}{Periods to import}

\item{date_insert}{Should the \code{date_insert} variable be returned?}

\item{spread}{Should the return be reshaped to have variables as column 
names? The \code{date_insert} and \code{count} variables are dropped so this
can occur.}

\item{tz}{What time-zone should the dates be returned as? Default is 
\code{"UTC"}.}
}
\value{
Data frame.
}
\description{
Function to import monthly summaries from an \strong{smonitor} database.
}
\author{
Stuart K. Grange
}
