% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_functions.R
\name{import_processes}
\alias{import_processes}
\title{Function to import process table from database.}
\usage{
import_processes(con, extra = TRUE)
}
\description{
Function to import process table from database.
}

