% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_processes_europe.R
\name{import_processes_europe}
\alias{import_processes_europe}
\title{Function to import \code{`processes`} table from the European 
\strong{smonitor} database.}
\usage{
import_processes_europe(con, country_code = NA, minimal = FALSE,
  print_query = FALSE)
}
\arguments{
\item{con}{A \strong{smonitor} database connection.}

\item{country_code}{A two digit ISO country code used for a regular
expression \code{WHERE} clause. This will usually be parsed from site codes. 
If \code{country_code} is \code{NA}, all ISO country codes in the database 
will be used.}

\item{minimal}{Should a bare-minimum SQL be run? This statment is small as 
possible to make it fast when used within other functions.}

\item{print_query}{Should the SQL query string be printed?}
}
\description{
\code{import_processes_europe} is an optimised version of 
\code{import_processes} which includes a regular expression filter to speed
up the retreval of processes because the table is larger than normal.
}
\author{
Stuart K. Grange
}
\seealso{
\code{\link{import_processes}}
}

