% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_simple_summaries.R
\name{import_simple_summaries}
\alias{import_simple_summaries}
\title{Function to import simple summaries from an \strong{smonitor} database.}
\usage{
import_simple_summaries(
  con,
  site = NA,
  summary = "month",
  date_insert = FALSE,
  tz = "UTC"
)
}
\arguments{
\item{con}{Database connection to an \strong{smonitor} database.}

\item{site}{A vector of sites to import. If not used, all sites are selected
and returned.}

\item{summary}{Summaries to import, can either be \code{"month"} or 
\code{"year"}.}

\item{date_insert}{Should the \code{date_insert} variable be imported?}

\item{tz}{What time zone should the dates be returned as?}
}
\value{
Tibble.
}
\description{
Function to import simple summaries from an \strong{smonitor} database.
}
\author{
Stuart K. Grange
}
