% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sites.R
\name{import_sites}
\alias{import_sites}
\title{Function to import \code{`sites`} table from a \strong{smonitor} database.}
\usage{
import_sites(con, extra = TRUE, print_query = FALSE)
}
\arguments{
\item{con}{Database connection.}

\item{extra}{Return extra data? Default is \code{TRUE}.}

\item{print_query}{Should the SQL query string be printed?}
}
\description{
Function to import \code{`sites`} table from a \strong{smonitor} database.
}
