% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_smonitor_details.R
\name{import_smonitor_details}
\alias{import_smonitor_details}
\title{Function to import data from an \strong{smonitor} \code{`smonitor_details`} 
table.}
\usage{
import_smonitor_details(con, tz = "UTC")
}
\arguments{
\item{con}{Database connection for an \strong{smonitor} database.}

\item{tz}{What time zone should the \code{date} variable be represented in?}
}
\value{
Tibble.
}
\description{
Function to import data from an \strong{smonitor} \code{`smonitor_details`} 
table.
}
\author{
Stuart K. Grange
}
