% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_functions.R
\name{import_source}
\alias{import_source}
\title{Functions to import source data from a database.}
\usage{
import_source(con, process)
}
\arguments{
\item{con}{Database connection.}

\item{process}{A process, an integer key.}
}
\description{
Functions to import source data from a database.
}
\author{
Stuart K. Grange
}

