% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_trend_summaries.R
\name{import_trend_summaries}
\alias{import_trend_summaries}
\title{Function to import trend summaries from an \strong{smonitor} database.}
\usage{
import_trend_summaries(
  con,
  site = NA,
  variable = NA,
  aggregation = "monthly_mean",
  date_insert = FALSE,
  spread = FALSE,
  tz = "UTC",
  print_query = FALSE
)
}
\arguments{
\item{con}{Database connection.}

\item{site}{Sites to import.}

\item{variable}{Variables to import.}

\item{aggregation}{Aggregation to import.}

\item{date_insert}{Should the \code{date_insert} variable be returned?}

\item{spread}{Should the return be reshaped to have variables as column 
names? The \code{date_insert} and \code{count} variables are dropped so this
can occur.}

\item{tz}{What time-zone should the dates be returned as? Default is 
\code{"UTC"}.}

\item{print_query}{Should the SQL query statement be printed?}
}
\value{
Data frame.
}
\description{
Function to import trend summaries from an \strong{smonitor} database.
}
\author{
Stuart K. Grange
}
