% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_functions.R
\name{import_uk}
\alias{import_uk}
\title{Function to import UK air quality data from a \strong{smonitor} database.}
\usage{
import_uk(con, site, start = 1970, end = NA, tz = "UTC", extra = FALSE,
  heathrow = FALSE)
}
\arguments{
\item{con}{Database connection.}

\item{site}{A site code such as \code{"my1"}.}

\item{start}{Start date to import.}

\item{end}{End date to import.}

\item{tz}{Time-zone for the dates to be parsed into. Default is \code{"UTC"}.}

\item{extra}{Should the data frame contain extra variables? Default is 
\code{FALSE}.}

\item{heathrow}{Should Heathrow's meteorological data be added to the data
frame? Useful for dealing with London's air quality data.}
}
\description{
Function to import UK air quality data from a \strong{smonitor} database.
}
\author{
Stuart K. Grange
}

