% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_zone_names.R
\name{import_zone_names}
\alias{import_zone_names}
\title{Function to import zone names from \code{`zones`} table in an \strong{smonitor}
database.}
\usage{
import_zone_names(con)
}
\arguments{
\item{con}{Database connection.}
}
\description{
Function to import zone names from \code{`zones`} table in an \strong{smonitor}
database.
}
\author{
Stuart K. Grange
}
