% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_zones.R
\name{import_zones}
\alias{import_zones}
\title{Function to import zones from an \strong{smonitor} database.}
\usage{
import_zones(con, zone = NA)
}
\arguments{
\item{con}{Database connection.}

\item{zone}{A zone vector. If not used, all zones will be returned.}
}
\value{
SpatialPolygonsDataFrame with WGS84 projection.
}
\description{
Function to import zones from an \strong{smonitor} database.
}
\seealso{
\code{\link{import_by_zone}}, \code{\link{import_zone_names}}
}
\author{
Stuart K. Grange
}
