% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_aurn_data.R
\name{insert_aurn_data}
\alias{insert_aurn_data}
\title{Function to get observations from \strong{openair's} \code{importAURN} 
function and insert them into a \strong{smonitor} database.}
\usage{
insert_aurn_data(con, site, start, end = NA, verbose = TRUE)
}
\arguments{
\item{con}{Database connection.}

\item{site}{Site code.}

\item{start}{Start year to download and insert.}

\item{end}{End year to download and insert.}

\item{verbose}{Should the function give messages and be chatty? Default is 
\code{TRUE}.}
}
\description{
Site-variable combinations need to be present in the database's process table,
otherwise they will be silently filtered and not be inserted.
}
\author{
Stuart K. Grange
}
