% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_kings_college_data.R
\name{insert_kings_college_data}
\alias{insert_kings_college_data}
\title{Function to get observations from \strong{openair's} \code{importKCL} 
function and insert them into a \strong{smonitor} database.}
\usage{
insert_kings_college_data(con, site, start, end = NA,
  data_source = "openair:importKCL", service = TRUE, verbose = TRUE)
}
\arguments{
\item{con}{Database connection.}

\item{site}{Site code.}

\item{start}{Start year to download and insert.}

\item{end}{End year to download and insert.}

\item{data_source}{\code{data_source} variable to use as a filter on the 
\code{`sites`} table. The default is \code{"openair:importKCL"}.}

\item{service}{Should a \code{service} variable be used to filter sites? 
If \code{TRUE}, the service value should be \code{1} to indicate service by
this function.}

\item{verbose}{Should the function give messages and be chatty? Default is 
\code{TRUE}.}
}
\description{
Site-variable combinations need to be present in the database's process table,
otherwise they will be silently filtered and not be inserted.
}
\author{
Stuart K. Grange
}

