% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_observations.R
\name{insert_observations}
\alias{insert_observations}
\title{Function to insert observations into the \code{`observations`} in a 
\strong{smonitor} database.}
\usage{
insert_observations(con, df, message = FALSE)
}
\arguments{
\item{con}{Database connection}

\item{df}{Data frame containing observations to be inserted.}

\item{message}{Should a message of counts and size be displayed?}
}
\description{
\code{insert_observations} will handle the variable/column order, the date 
conversions, add a \code{"date_insert"} variable, and the database insert for 
a \strong{smonitor} database.
}
\examples{
\dontrun{

# Insert some data
insert_observations(con, data_test)

# Insert some data, with a message
insert_observations(con, data_test, message = TRUE)

}

}
\author{
Stuart K. Grange
}

