% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_simple_summaries.R
\name{insert_simple_summaries}
\alias{insert_simple_summaries}
\title{Function to insert simple summaries into a \strong{smonitor} database.}
\usage{
insert_simple_summaries(con, df, .progress = FALSE)
}
\arguments{
\item{con}{Database connection to a \strong{smonitor} database.}

\item{df}{Tibble from \code{calculate_simple_summaries}.}

\item{.progress}{Should a progress bar be displayed?}
}
\value{
Invisible \code{con}.
}
\description{
\code{insert_simple_summaries} will delete old summaries before inserting.
}
\author{
Stuart K. Grange.
}
