% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrity_check.R
\name{integrity_check}
\alias{integrity_check}
\title{Function to test data-model integrity of an \strong{smonitor} database.}
\usage{
integrity_check(con)
}
\arguments{
\item{con}{Database connection.}
}
\description{
\code{integrity_check} tests if the foreign keys are accounted for in the
\strong{smonitor} tables.
}
\author{
Stuart K. Grange
}

