% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invalidate_bad_wind_direction.R
\name{invalidate_bad_wind_direction}
\alias{invalidate_bad_wind_direction}
\title{Function to invalidate wind direction processes which are less than \code{0}
or greater than \code{360} degrees in an \code{smonitor} database.}
\usage{
invalidate_bad_wind_direction(con, print_query = FALSE)
}
\arguments{
\item{con}{Database connection.}

\item{print_query}{Should the SQL queries be printed?}
}
\value{
Invisible.
}
\description{
Function to invalidate wind direction processes which are less than \code{0}
or greater than \code{360} degrees in an \code{smonitor} database.
}
\author{
Stuart K. Grange
}
