% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz_tz.R
\name{nz_tz}
\alias{nz_tz}
\title{Function for New Zealand's "psudo time-zone" string used for air quality 
monitoring.}
\usage{
nz_tz()
}
\description{
The time-zone string is \code{"Etc/GMT-12"}.
}
\author{
Stuart K. Grange
}
