% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_plot.R
\name{polar_plot}
\alias{polar_plot}
\title{Function to plot polar plots.}
\usage{
polar_plot(df, variable, ws = "ws", wd = "wd", polar = TRUE,
  contour = FALSE)
}
\description{
\code{polar_plot} is slow due to the combination of using \strong{ggplot}'s 
\code{geom_tile} and \code{coord_polar} at the same time.
}
\author{
Stuart K. Grange
}

