% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_observations_table.R
\name{prepare_observations_table}
\alias{prepare_observations_table}
\title{Function to prepare a data frame for insert into the \code{`observations`} 
table in a \strong{smonitor} database.}
\usage{
prepare_observations_table(
  df,
  drop = FALSE,
  convert = FALSE,
  as_tibble = FALSE,
  include_date_insert = TRUE
)
}
\arguments{
\item{df}{Data frame or tibble.}

\item{drop}{Should variables not contained in an \code{`observations`} table 
be dropped?}

\item{convert}{Should the variables be coerced into the correct data types?}

\item{as_tibble}{Should the return be a tibble?}

\item{include_date_insert}{Should the \code{date_insert} variable be included?}
}
\value{
Data frame or tibble.
}
\description{
Function to prepare a data frame for insert into the \code{`observations`} 
table in a \strong{smonitor} database.
}
\author{
Stuart K. Grange
}
