% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_invalid_values.R
\name{set_invalid_values}
\alias{set_invalid_values}
\title{Function to set invalid values to missing (\code{NA}) based on 
\strong{smonitor}'s convention for validity.}
\usage{
set_invalid_values(value, validity)
}
\arguments{
\item{value}{A numeric vector representing values (usually 
measurements/observations).}

\item{validity}{An integer vector representing validity.}
}
\description{
Validity integers less-than and equal to zero are considered invalid and will
be set to missing and validity integers that are either missing or positive 
are considered valid.
}
\author{
Stuart K. Grange
}
