% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smonitor_details.R
\name{smonitor_details}
\alias{smonitor_details}
\title{Function to get details for a \strong{smonitor} database.}
\usage{
smonitor_details(con, task = NA_character_)
}
\arguments{
\item{con}{Database connection for a \strong{smonitor} database.}

\item{task}{A character vector to indicate what task is being (or has been) 
done.}
}
\value{
Tibble.
}
\description{
Function to get details for a \strong{smonitor} database.
}
\author{
Stuart K. Grange
}
