% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smonitor_indices_drop.R
\name{smonitor_indices_drop}
\alias{smonitor_indices_create}
\alias{smonitor_indices_drop}
\title{Function to drop indices from an smonitor database, usually used before a 
batch insert.}
\usage{
smonitor_indices_drop(con)

smonitor_indices_create(con)
}
\arguments{
\item{con}{Database connection.}

\item{con}{Database connection.}
}
\description{
Function to drop indices from an smonitor database, usually used before a 
batch insert.

Function to create indices from an smonitor database, usually used after a 
batch insert.
}
\author{
Stuart K. Grange

Stuart K. Grange
}

