% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smonitor_update_tasks.R
\name{smonitor_update_tasks}
\alias{smonitor_update_tasks}
\title{Function to update things in an \strong{smonitor} database.}
\usage{
smonitor_update_tasks(con, tz = "UTC")
}
\arguments{
\item{con}{Database connection.}

\item{tz}{Time-zone for dates.}
}
\description{
\code{update_smonitor_things} will update variables in the \code{`processes`}, 
\code{`sites`}, and \code{`row_counts`} tables and is intended to be run as
as scheduled task.
}
\author{
Stuart K. Grange
}
