% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_smonitor_integrity.R
\name{test_smonitor_integrity}
\alias{test_smonitor_integrity}
\title{Function to test the integrity of a \strong{smonitor} database.}
\usage{
test_smonitor_integrity(con)
}
\arguments{
\item{con}{Database connection.}
}
\value{
Tibble.
}
\description{
Function to test the integrity of a \strong{smonitor} database.
}
\author{
Stuart K. Grange
}
