% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theil_sen_trend_test.R
\name{theil_sen_trend_test}
\alias{theil_sen_trend_test}
\title{Function to test a time series with the Theil-Sen estimator.}
\usage{
theil_sen_trend_test(
  df,
  variable = "value",
  deseason = FALSE,
  alpha = 0.05,
  auto_correlation = FALSE,
  period = "month"
)
}
\arguments{
\item{df}{Input data frame, containing time series observations.}

\item{variable}{Variable name to test.}

\item{deseason}{Should the time series be deseaonsalised before the trend 
test is conducted?}

\item{alpha}{Confidence interval of the slope, default is 0.05 for 95 \%
confidence intervals.}

\item{auto_correlation}{Should auto correlation be considered in the 
estimates?}

\item{period}{Period of input time series. Default is \code{"month"} but can
also be \code{"year"}.}
}
\value{
Tibble with one observation/row.
}
\description{
Function to test a time series with the Theil-Sen estimator.
}
\seealso{
\code{\link[openair]{TheilSen}}
}
\author{
Stuart K. Grange.
}
