% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_date_span_variables.R
\name{update_date_span_variables}
\alias{update_date_span_variables}
\title{Function to update the \code{date_start}, \code{date_end}, variables in the
\code{`processes`} and \code{`sites`} tables in a \strong{smonitor} database.}
\usage{
update_date_span_variables(con, tz = "UTC", verbose = FALSE)
}
\arguments{
\item{con}{Database connection.}

\item{tz}{Time-zone Time-zone for the dates to be represented as.}

\item{progress}{Type of progress bar to display.}
}
\description{
\code{update_date_span_variables} will also insert (or replace) the 
\code{`row_counts`} table.
}
\author{
Stuart K. Grange
}
