% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_row_counts.R
\name{update_row_counts}
\alias{update_row_counts}
\title{Function to get row counts and insert as a table in a database.}
\usage{
update_row_counts(con, estimate = FALSE, read = FALSE)
}
\arguments{
\item{con}{Database connection.}

\item{estimate}{Should the row counts be estimated rather than counted? Only
works for PostgreSQL databases.}

\item{read}{Should the table be immediately read after being updated? Default
is \code{FALSE}.}
}
\description{
Function to get row counts and insert as a table in a database.
}
\author{
Stuart K. Grange
}
\seealso{
\code{\link{db_count_rows}}
}

