% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_site_span_variables.R
\name{update_site_span_variables}
\alias{update_site_span_variables}
\title{Function to update \code{date_start} and \code{date_end}, for sites and then
update the \code{`sites`} table for an \strong{smonitor} database.}
\usage{
update_site_span_variables(con, tz = "UTC", progress = "none")
}
\arguments{
\item{con}{Database connection.}

\item{tz}{Time-zone to parse the dates to.}

\item{progress}{Type of progress bar to display for the update statements. 
Default is \code{"none"}.}
}
\description{
Use \code{update_site_span_variables} after 
\code{\link{update_process_span_variables}} because it queries the 
\code{`processes`} table rather than \code{`observations`}.
}
\seealso{
\code{\link{update_process_span_variables}}
}

