% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_site_spans.R
\name{update_site_spans}
\alias{update_site_spans}
\title{Function to update the \code{date_start}, \code{date_end}, 
\code{observation_count}, and \code{variables_monitored} variables for a 
\code{`sites`} table in an smonitor database.}
\usage{
update_site_spans(con, site = NA, variables_monitored = FALSE)
}
\arguments{
\item{con}{Database connection.}

\item{site}{Vector of sites to update.}

\item{variables_monitored}{Should the \code{variables_monitored} variable 
also be updated?}
}
\value{
Invisible \code{con}.
}
\description{
Use \code{update_site_span} after \code{\link{update_process_spans}} because 
it queries the \code{`processes`} table rather than \code{`observations`}.
}
\seealso{
\code{\link{update_process_spans}}, 
\code{\link{update_date_span_variables}}
}
\author{
Stuart K. Grange
}
