% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_validity.R
\name{update_validity}
\alias{update_validity}
\title{Function to update validity variable in a \code{`observations`} table in a 
\strong{smonitor} database.}
\usage{
update_validity(con, process, summary = 0, tz = "UTC",
  verbose = FALSE)
}
\arguments{
\item{con}{Database connection.}

\item{process}{A vector of processes.}

\item{summary}{Summary to invalidate. This will usually be \code{0} for source
data and can only take one value.}

\item{tz}{Time zone to conduct invalidations in.}

\item{verbose}{Should the function give messages?}
}
\value{
Invisible \code{con}.
}
\description{
\code{update_validity} does not update the validity variable with SQL, it 
queries the database, re-tests, deletes the old observations and inserts the
re-tested table.
}
\author{
Stuart K. Grange
}
