% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_variables_monitored.R
\name{update_variables_monitored}
\alias{update_variables_monitored}
\title{Function to update the \code{variables_monitored} variable in the 
\code{`site`} table of an \strong{smonitor} database.}
\usage{
update_variables_monitored(con, site = NA)
}
\arguments{
\item{con}{Database connection.}

\item{site}{Vector of sites to update.}
}
\value{
Invisible \code{con}.
}
\description{
Function to update the \code{variables_monitored} variable in the 
\code{`site`} table of an \strong{smonitor} database.
}
\author{
Stuart K. Grange
}
