% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity_test.R
\name{validity_test}
\alias{validity_test}
\title{Function to add a validity variable to a data frame based on date ranges.}
\usage{
validity_test(df, df_look)
}
\description{
Function to add a validity variable to a data frame based on date ranges.
}
\author{
Stuart K. Grange.
}

