% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_rose.R
\name{wind_rose}
\alias{wind_rose}
\title{Function to plot wind roses.}
\usage{
wind_rose(data, spd = "ws", dir = "wd", spdres = 2, dirres = 30,
  spdmin = 2, spdmax = 20, spdseq = NULL, palette = "YlGnBu",
  countmax = NA, debug = 0)
}
\arguments{
\item{data}{Data frame.}

\item{spd}{Wind speed variable. Default is \code{"ws"}.}

\item{dir}{Wind direction variable. Default is \code{"wd"}.}

\item{spdres}{Wind speed resolution. Default is \code{2}.}

\item{dirres}{Wind direction resolution. Default is \code{30}.}

\item{spdmin}{Wind speed minimum. Default is \code{2}.}

\item{spdmax}{Wind speed maximum. Default is \code{20}.}

\item{spdseq}{Wind speed sequence. An override.}

\item{palette}{Colour palette to use. Default is \code{"YlGnBu"}.}

\item{countmax}{An adjustment override for the wind speed scale.}

\item{debug}{Show debug messages. Use \code{0} or \code{1}, not logicals.}
}
\description{
\code{wind_rose} uses \strong{ggplot2} and is a stacked bar chart with
polar coordinates. \code{wind_rose} bins wind speed and direction then 
aggregates to find counts.
}
\seealso{
\code{\link{windRose}} \href{http://stackoverflow.com/questions/17266780/wind-rose-with-ggplot-r}{stackoverflow.com}
}
\author{
Andy Clifton with a few changes by Stuart K. Grange
}
