% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_summaries.R
\name{calculate_summaries}
\alias{calculate_summaries}
\title{Function to calculate various time-series summaries.}
\usage{
calculate_summaries(con, df_map, start, end, insert = FALSE)
}
\arguments{
\item{con}{Database connection.}

\item{df_map}{A mapping table containing \code{"process"} and \code{"summary"}
variables.}

\item{start}{Start date for summaries.}

\item{end}{End date for summaries.}

\item{insert}{Should the data be inserted?}
}
\description{
\code{calculate_summaries} will correctly process wind speed and direction
with the help of \strong{openair}.
}
\author{
Stuart K. Grange
}

