% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_summaries.R
\name{calculate_summaries}
\alias{calculate_summaries}
\title{Function to calculate various time-series summaries.}
\usage{
calculate_summaries(con, df_map, insert = FALSE)
}
\arguments{
\item{con}{Database connection.}

\item{df_map}{A mapping table.}

\item{insert}{Should the data be inserted?}
}
\description{
Function to calculate various time-series summaries.
}
\author{
Stuart K. Grange
}

