% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_summaries.R
\name{calculate_summaries}
\alias{calculate_summaries}
\title{Function to calculate various time-series summaries for data stored in a 
\strong{smonitor} database.}
\usage{
calculate_summaries(con, df_map, start, end, tz = "UTC", insert = TRUE)
}
\arguments{
\item{con}{Database connection.}

\item{df_map}{A mapping table containing \code{"process"} and \code{"summary"}
variables.}

\item{start}{Start date for summaries.}

\item{end}{End date for summaries.}

\item{insert}{Should the data be inserted? Default is \code{TRUE}.}
}
\description{
\code{calculate_summaries} will handle aggregation methods, transformation,
and updating processes when the \strong{smonitor} databases are complete.
\code{calculate_summaries} will also correctly aggregate wind direction.
}
\examples{
\dontrun{

# ... create database connection
# ... create look-up table

# Calculate all summaries in data_look_up for 2014 and insert them in the 
# database
calculate_summaries(con, data_look_up, start = 2014, start = 2014)

}

}
\author{
Stuart K. Grange
}
\seealso{
\code{\link{timeAverage}}, \code{\link{dplyr::summarise}}
}

