% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_observations.R
\name{delete_observations}
\alias{delete_observations}
\title{Function to delete observations from the \code{`observations`} table in a
\strong{smonitor} database.}
\usage{
delete_observations(con, df, groups = c("process", "summary"),
  match = "between", verbose = FALSE, progress = "none")
}
\arguments{
\item{con}{Database connection.}

\item{df}{Data frame; usually an object which is about to be inserted into 
a database table. \code{delete_observations} uses this as a mapping table to
know what to send the database in the delete statement.}

\item{groups}{Groups in \code{df} to apply the deleting function to. Default 
is \code{c("process", "summary")} and should not need changing.}

\item{match}{Type of match to use for dates. \code{"between"} and 
\code{"all"} are supported. Beware that \code{"all"} will remove all 
observations associated with a process.}

\item{verbose}{Should the function give messages?}

\item{progress}{Type of progress bar to display.}
}
\value{
Invisible. A database delete.
}
\description{
\code{delete_observations} is generally used immediately before a database 
insert so old observations are deleted before a new set of observations are 
inserted. This process avoids duplicate observations which are almost always
undesirable.
}
\details{
\code{delete_observations} allows the use of groups based on the input data
frame.
}
\seealso{
\code{\link{db_execute}}, \code{\link{insert_observations}}
}
\author{
Stuart K. Grange
}
