% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_observations.R
\name{delete_observations}
\alias{delete_observations}
\title{Function to delete observations from the \code{`observations`} table in a
\strong{smonitor} database.}
\usage{
delete_observations(con, df, match = "between")
}
\arguments{
\item{con}{Database connection.}

\item{df}{Data frame; usually an object which is about to be inserted into 
a database table.}

\item{match}{Type of match to use for dates. Currently only \code{"between"}
is supported.}
}
\description{
\code{delete_observations} is generally used immediately before a database 
insert so old observations are deleted before a new set of observations are 
inserted. This process avoids duplicate observations which are almost always
undesirable.
}
\details{
\code{delete_observations} does not currently handle groups, but will in the 
future. It is recommended that \code{plyr::a_ply} is used for this task.
}
\author{
Stuart K. Grange
}
\seealso{
\link{db_send}
}

