% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_process.R
\name{delete_process}
\alias{delete_process}
\title{Function to delete all observations for a process from the 
\code{`observations`} table in a \strong{smonitor} database.}
\usage{
delete_process(con, process)
}
\arguments{
\item{con}{Database connection.}

\item{process}{A vector of processes.}
}
\description{
\code{delete_process} is usually used when a process needs to be purged and 
will delete all summaries associated with a process too. Use with care.
}
\author{
Stuart K. Grange
}

