% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_next_process_key.R
\name{get_next_process_key}
\alias{get_next_process_key}
\title{Function to get next \code{process} key for a \strong{smonitor} database.}
\usage{
get_next_process_key(con)
}
\arguments{
\item{con}{Database connection.}
}
\value{
Interger with length of 1.
}
\description{
Function to get next \code{process} key for a \strong{smonitor} database.
}
\author{
Stuart K. Grange
}
