% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_aggregations.R
\name{import_aggregations}
\alias{import_aggregations}
\title{Function to import \code{`aggregations`} table from a \strong{smonitor} 
database.}
\usage{
import_aggregations(con)
}
\arguments{
\item{con}{A \strong{smonitor} database connection.}
}
\value{
Tibble.
}
\description{
Function to import \code{`aggregations`} table from a \strong{smonitor} 
database.
}
\examples{
\dontrun{

# Import aggregations from a smonitor database
data_aggregations <- import_aggregations(con)

}

}
\author{
Stuart K. Grange
}
