% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_functions.R
\name{import_any}
\alias{import_any}
\title{Function to import any observational data from a \strong{smonitor} database.}
\usage{
import_any(con, process, summary, start = 1970, end = NA, tz = "UTC",
  valid = FALSE, extra = TRUE)
}
\arguments{
\item{con}{Database connection.}

\item{process}{A process, an integer key.}

\item{summary}{A summary, an integer key.}

\item{start}{Start date to import.}

\item{end}{End date to import.}

\item{tz}{Time-zone for the dates to be parsed into. Default is \code{"UTC"}.}

\item{valid}{Should invalid observations be filtered out? Default is 
\code{FALSE}.}
}
\description{
Function to import any observational data from a \strong{smonitor} database.
}
\author{
Stuart K. Grange
}

