% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_any.R
\name{import_any}
\alias{import_any}
\title{Function to import observational data from a \strong{smonitor} database.}
\usage{
import_any(con, process, summary = NA, start = 1970, end = NA,
  tz = "UTC", valid_only = TRUE)
}
\arguments{
\item{con}{A \strong{smonitor} database connection.}

\item{process}{A process, an integer key.}

\item{summary}{A summary, an integer key. If summary is \code{NA}, then only 
the \code{process} will be used in the \code{WHERE} clause.}

\item{start}{Start date to import.}

\item{end}{End date to import.}

\item{tz}{Time-zone for the dates to be parsed into. Default is \code{"UTC"}.}

\item{valid_only}{Should invalid observations be filtered out? Default is 
\code{TRUE}.}
}
\description{
Function to import observational data from a \strong{smonitor} database.
}
\author{
Stuart K. Grange
}

