% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_by_site.R
\name{import_by_site}
\alias{import_by_site}
\title{Function to import data from an \code{smonitor} database based on sites.}
\usage{
import_by_site(con, site, start = 1970, end = NA, period = "hour",
  valid = TRUE, pad = TRUE, tz = "UTC", extra = TRUE)
}
\arguments{
\item{con}{Database connection.}

\item{site}{A site code such as \code{"my1"}.}

\item{start}{Start date to import.}

\item{end}{End date to import.}

\item{period}{Averaging period.}

\item{valid}{Should only valid data be returned?}

\item{pad}{Should the time-series be padded to ensure all dates in the 
observation period are present?}

\item{tz}{Time-zone for the dates to be parsed into. Default is \code{"UTC"}.}

\item{extra}{Should the data frame contain extra variables? Default is 
\code{FALSE}.}
}
\description{
Function to import data from an \code{smonitor} database based on sites.
}
\author{
Stuart K. Grange
}

