% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_by_site.R
\name{import_by_site}
\alias{import_by_site}
\title{Function to import data from an \strong{smonitor} database based on sites.}
\usage{
import_by_site(con, site, variable = NA, start = 1970, end = NA,
  period = "hour", valid_only = TRUE, pad = TRUE, tz = "UTC",
  spread = FALSE, europe = FALSE, date_end = TRUE, date_insert = FALSE,
  site_name = TRUE, print_query = FALSE)
}
\arguments{
\item{con}{A \strong{smonitor} database connection.}

\item{site}{A site code such as \code{"my1"}.}

\item{variable}{An optional variable vector. If not used, all variables will
be selected and returned.}

\item{start}{Start date to import.}

\item{end}{End date to import.}

\item{period}{Averaging period. Default is \code{"hour"}. \code{period} can
also take the value \code{"any"} which will return all periods, but 
\code{pad} argument will be ignored.}

\item{valid_only}{Should only valid data be returned? Default is \code{TRUE}.}

\item{pad}{Should the time-series be padded to ensure all dates in the 
observation period are present? Default is \code{TRUE}.}

\item{tz}{Time-zone for the dates to be parsed into. Default is \code{"UTC"}.}

\item{spread}{Should the data frame take the wider format resulting from
spreading the data? Default is \code{FALSE}.}

\item{europe}{Should \code{import_processes_europe} be used rather than
\code{import_processes}?}

\item{date_end}{Should the return include the \code{date_end} variable? 
Default is \code{TRUE}.}

\item{date_insert}{Should the return include the \code{date_insert} variable? 
Default is \code{FALSE}.}

\item{site_name}{Should the return include the \code{site_name} variable? 
Default is \code{TRUE}.}

\item{print_query}{Should the SQL query string be printed?}
}
\description{
\code{import_by_site} is considered the primary user-focused importing 
function while \code{import_any} is the primary lower-level function.
}
\seealso{
\code{\link{import_any}}
}
\author{
Stuart K. Grange
}
