% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_by_site.R
\name{import_by_site}
\alias{import_by_site}
\title{Function to import data from an \strong{smonitor} database based on sites.}
\usage{
import_by_site(con, site = NA, variable = NA, start = 1970,
  end = NA, period = "hour", include_sums = TRUE,
  valid_only = TRUE, pad = TRUE, tz = "UTC", spread = FALSE,
  date_end = TRUE, date_insert = FALSE, site_name = TRUE,
  unit = TRUE, print_query = FALSE)
}
\arguments{
\item{con}{A \strong{smonitor} database connection.}

\item{site}{A vector of sites.}

\item{variable}{An optional variable vector. If not used, all variables will
be returned.}

\item{start}{Start date to import.}

\item{end}{End date to import.}

\item{period}{Averaging period. Default is \code{"hour"}. \code{period} can
also take the value \code{NA} which will return all periods, but \code{pad} 
argument will be ignored.}

\item{include_sums}{Should processes with the appropriate \code{period} which 
have been summarised/aggregated with a sum function also be returned? This is
useful for variables such as rainfall, snow, and precipitation where normally
the aggregations do not take the form of the mean. Default is \code{TRUE}.}

\item{valid_only}{Should only valid data be returned? Default is \code{TRUE}.}

\item{pad}{Should the time-series be padded to ensure all dates in the 
observation period are present? Default is \code{TRUE}.}

\item{tz}{Time-zone for the dates to be parsed into. Default is \code{"UTC"}.}

\item{spread}{Should the data frame take the wider format resulting from
spreading the data? Default is \code{FALSE}.}

\item{date_end}{Should the return include the \code{date_end} variable? 
Default is \code{TRUE}.}

\item{date_insert}{Should the return include the \code{date_insert} variable? 
Default is \code{FALSE}.}

\item{site_name}{Should the return include the \code{site_name} variable? 
Default is \code{TRUE}.}

\item{unit}{Should the processes' units be included in the return?}

\item{print_query}{Should the SQL query string be printed?}
}
\description{
\code{import_by_site} is considered the primary user-focused importing 
function while \code{import_by_process} is the primary lower-level function.
}
\examples{
\dontrun{

# Get hourly means, the default
data_site_hour <- import_by_site(con, site = "site_005", start = 2010)

# Get hourly means and spread the data, useful for some analyses
data_site_hour_spread <- import_by_site(
  con, 
  site = "site_005", 
  start = 2010,
  spread = TRUE
)

# Get daily means for two sites
data_site_hour <- import_by_site(
  con, 
  site = c("site_005", "site_100"), 
  period = "day", 
  start = 2010,
  spread = TRUE
)

}

}
\seealso{
\code{\link{import_by_process}}
}
\author{
Stuart K. Grange
}
