% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_functions.R
\name{import_hourly_means}
\alias{import_hourly_means}
\title{Function to import hourly means from a \strong{smonitor} database.}
\usage{
import_hourly_means(con, process, start = 1970, end = NA, tz = "UTC")
}
\arguments{
\item{con}{Database connection.}

\item{process}{A process, an integer key.}

\item{start}{Start date to import.}

\item{end}{End date to import.}

\item{tz}{Time-zone for the dates to be parsed into. Default is \code{"UTC"}.}
}
\description{
Function to import hourly means from a \strong{smonitor} database.
}
\author{
Stuart K. Grange
}

