% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_invalidations.R
\name{import_invalidations}
\alias{import_invalidations}
\title{Function to import \code{`invalidations`} table from a \strong{smonitor} 
database.}
\usage{
import_invalidations(con, tz = "UTC")
}
\arguments{
\item{con}{Database connection.}

\item{tz}{What time-zone should the dates be in? Default is \code{"UTC"}.}
}
\description{
Function to import \code{`invalidations`} table from a \strong{smonitor} 
database.
}
