% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_functions.R
\name{import_nz}
\alias{import_nz}
\title{Function to import data from an air quality monitoring data with the 
\strong{smonitor} data model.}
\usage{
import_nz(con, process, summary, start = 1970, end = NA,
  tz = "Etc/GMT-12", extra = TRUE)
}
\arguments{
\item{con}{Database connection.}

\item{process}{A process, an integer key.}

\item{summary}{A summary, an integer key.}

\item{start}{Start date to import.}

\item{end}{End date to import.}

\item{tz}{Time-zone for the dates to be parsed into. Default is 
\code{"Etc/GMT-12"}.}
}
\description{
Function to import data from an air quality monitoring data with the 
\strong{smonitor} data model.
}
\author{
Stuart K. Grange
}

