% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sites.R
\name{import_sites}
\alias{import_sites}
\title{Function to import \code{`sites`} table from a \strong{smonitor} database.}
\usage{
import_sites(con, tz = "UTC", correct_integers = TRUE)
}
\arguments{
\item{con}{Database connection.}

\item{tz}{What time zone should the \code{date_start} and \code{date_end}
variables be represented as?}

\item{correct_integers}{Should integers be corrected for a SQLite issue?}
}
\value{
Tibble.
}
\description{
Function to import \code{`sites`} table from a \strong{smonitor} database.
}
\examples{
\dontrun{

# Import sites from a smonitor database
data_sites <- import_sites(con)

}

}
\author{
Stuart K. Grange
}
