% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sites.R
\name{import_sites}
\alias{import_sites}
\title{Function to import \code{`sites`} table from a \strong{smonitor} database.}
\usage{
import_sites(con, tz = "UTC", print_query = FALSE)
}
\arguments{
\item{con}{Database connection.}

\item{tz}{What time zone should the \code{date_start} and \code{date_end}
variables be represented as?}

\item{print_query}{Should the SQL query be printed?}
}
\value{
Data frame.
}
\description{
Function to import \code{`sites`} table from a \strong{smonitor} database.
}
\author{
Stuart K. Grange
}
