% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_functions.R
\name{import_sites}
\alias{import_sites}
\title{Function to import sites table from a \strong{smonitor} database.}
\usage{
import_sites(con, extra = TRUE)
}
\arguments{
\item{con}{Database connection.}

\item{extra}{Return extra data? Default is \code{TRUE}.}
}
\description{
Function to import sites table from a \strong{smonitor} database.
}

