% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_functions.R
\name{import_summaries}
\alias{import_summaries}
\title{Function to import aggregation table from database.}
\usage{
import_summaries(con, extra = TRUE)
}
\description{
Function to import aggregation table from database.
}

