% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_summaries.R
\name{import_summaries}
\alias{import_summaries}
\title{Function to import \code{`summaries`} table from a \strong{smonitor} database.}
\usage{
import_summaries(con, extra = TRUE)
}
\arguments{
\item{con}{Database connection.}

\item{extra}{Return extra data? Default is \code{TRUE}.}
}
\description{
Function to import \code{`summaries`} table from a \strong{smonitor} database.
}
