% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_summaries.R
\name{import_summaries}
\alias{import_summaries}
\title{Function to import \code{`summaries`} table from a \strong{smonitor} database.}
\usage{
import_summaries(con)
}
\arguments{
\item{con}{Database connection.}
}
\value{
Data frame.
}
\description{
Function to import \code{`summaries`} table from a \strong{smonitor} database.
}
\examples{
\dontrun{

# Import summaries from a smonitor database
data_summaries <- import_summaries(con)

}

}
\author{
Stuart K. Grange
}
