% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_trend_tests.R
\name{import_trend_tests}
\alias{import_trend_tests}
\title{Function to import the \code{`trend_test`} table from an \strong{smonitor} 
database.}
\usage{
import_trend_tests(con)
}
\arguments{
\item{con}{Database connection.}
}
\value{
Data frame.
}
\description{
Function to import the \code{`trend_test`} table from an \strong{smonitor} 
database.
}
\author{
Stuart K. Grange
}
