% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_uk.R
\name{import_uk}
\alias{import_uk}
\title{Function to import UK air quality data from a \strong{smonitor} database.}
\usage{
import_uk(con, site, start = 1970, end = NA, tz = "UTC", extra = FALSE,
  heathrow = FALSE)
}
\arguments{
\item{con}{Database connection.}

\item{site}{A site code such as \code{"my1"}.}

\item{start}{What is the start date of data to be returned? Ideally, the 
date format should be \code{yyyy-mm-dd}, but the UK locale convention of 
\code{dd/mm/yyyy} will also work. Years as strings or integers work too and
will floor-rounded.}

\item{end}{What is the end date of data to be returned? Ideally, the 
date format should be \code{yyyy-mm-dd}, but the UK locale convention of 
\code{dd/mm/yyyy} will also work. Years as strings or integers work too and 
will be ceiling-rounded.}

\item{tz}{Time-zone for the dates to be parsed into. Default is \code{"UTC"}.}

\item{extra}{Should the data frame contain extra variables? Default is 
\code{FALSE}.}

\item{heathrow}{Should Heathrow's meteorological data be added to the data
frame? Useful for dealing with London's air quality data.}
}
\description{
Function to import UK air quality data from a \strong{smonitor} database.
}
\author{
Stuart K. Grange
}

