% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_observations.R
\name{insert_observations}
\alias{insert_observations}
\title{Function to insert observations into \code{`observations`} table for a 
\strong{smonitor} database.}
\usage{
insert_observations(con, df)
}
\arguments{
\item{con}{Database connection}

\item{df}{Data frame to be inserted.}
}
\description{
\code{insert_observations} will handle the variable order, the date 
conversions, add a \code{"date_insert"} variable, and the insert for a 
\strong{smonitor} database.
}
\examples{
\dontrun{

# Insert some data
insert_observations(con, data_test)

}

}
\author{
Stuart K. Grange
}

