% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_observations.R
\name{insert_observations}
\alias{insert_observations}
\title{Function to insert observations into the \code{`observations`} in a 
\strong{smonitor} database.}
\usage{
insert_observations(con, df, verbose = FALSE)
}
\arguments{
\item{con}{Database connection.}

\item{df}{Data frame containing observations to be inserted.}

\item{verbose}{Should the function give messages?}
}
\value{
Invisible, an insert into the \code{`observations`} database table.
}
\description{
\code{insert_observations} will handle the variable/column order, the date 
conversions, add a \code{"date_insert"} variable, and the database insert for 
a \strong{smonitor} database. A number of tests are conducted before insertion
is attempted.
}
\examples{
\dontrun{

# Insert some data
insert_observations(con, data_test)

# Insert some data, with some messages
insert_observations(con, data_test, verbose = TRUE)

}

}
\author{
Stuart K. Grange.
}
