% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_observations.R
\name{insert_observations}
\alias{insert_observations}
\title{Function to insert observations into the \code{`observations`} in a 
\strong{smonitor} database.}
\usage{
insert_observations(
  con,
  df,
  check_processes = TRUE,
  check_validity = TRUE,
  batch_size = NA,
  verbose = FALSE,
  progress = FALSE
)
}
\arguments{
\item{con}{Database connection.}

\item{df}{Data frame containing observations to be inserted.}

\item{check_processes}{Should the processes in \code{df} be tested for their
existence in the `processes` table before insert?}

\item{check_validity}{Should the validity in \code{df} be tested before 
insert?}

\item{batch_size}{Number of rows to insert per batch. Useful for large 
inserts.}

\item{verbose}{Should the function give messages?}

\item{progress}{If \code{batch_size} is not \code{NA}, should a progress bar 
be displayed?}
}
\value{
Invisible \code{con}.
}
\description{
\code{insert_observations} will handle the variable/column order, the date 
conversions, add a \code{"date_insert"} variable, and the database insert for 
a \strong{smonitor} database. A number of tests are conducted before insertion
is attempted.
}
\examples{
\dontrun{

# Insert observations data
insert_observations(con, data_test)

# Insert observations, this time with some messages
insert_observations(con, data_test, verbose = TRUE)

}

}
\seealso{
\code{\link{delete_observations}}
}
\author{
Stuart K. Grange.
}
