% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_wunderground_data.R
\name{insert_wunderground_data}
\alias{insert_wunderground_data}
\title{Function to get observations from wunderground and insert them into a 
\strong{smonitor} database.}
\usage{
insert_wunderground_data(con, site, start, end = NA, validity = FALSE,
  verbose = FALSE)
}
\arguments{
\item{con}{Database connection.}

\item{site}{Wunderground site code.}

\item{start}{Start date to download and insert.}

\item{end}{End date to download and insert.}

\item{validity}{Should the validity variable be updated? Not implemented yet.}

\item{verbose}{Should the function give messages? Default is \code{FALSE}.}
}
\description{
Site-variable combinations need to be present in the database's process table,
otherwise they will be silently filtered and not be inserted. New, downloaded
observations will take priority over those stored in the database and old 
observations are deleted with this function.
}
\author{
Stuart K. Grange
}

