% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tz_nz.R
\name{tz_nz}
\alias{tz_nz}
\alias{tz_central}
\alias{tz_eastern}
\title{Functions for time zone strings.}
\usage{
tz_nz()

tz_central()

tz_eastern()
}
\value{
Character vector length of 1.
}
\description{
Functions for time zone strings.
}
\author{
Stuart K. Grange
}
