% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_row_counts.R
\name{update_row_counts}
\alias{update_row_counts}
\title{Function to get row counts and insert as a table in a database.}
\usage{
update_row_counts(con, estimate = FALSE)
}
\arguments{
\item{con}{Database connection.}

\item{estimate}{Should the row counts be estimated rather than counted? Only
works for PostgreSQL databases.}
}
\value{
Invisible, tibble.
}
\description{
Function to get row counts and insert as a table in a database.
}
\seealso{
\code{\link[databaser]{db_count_rows}}
}
\author{
Stuart K. Grange
}
