% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_site_spans.R
\name{update_site_spans}
\alias{update_site_spans}
\title{Function to update the \code{date_start}, \code{date_end}, 
\code{observation_count}, \code{variables_monitored}, and \code{"data_source"}
 variables for a \code{`sites`} table in an \strong{smonitor} database.}
\usage{
update_site_spans(con, site = NA)
}
\arguments{
\item{con}{\strong{smonitor} database connection.}

\item{site}{Vector of sites to update.}
}
\value{
Invisible \code{con}.
}
\description{
Use \code{update_site_span} after \code{\link{update_process_spans}} because 
it queries the \code{`processes`} table rather than \code{`observations`}.
}
\seealso{
\code{\link{update_process_spans}}, 
\code{\link{update_date_span_variables}}
}
\author{
Stuart K. Grange
}
