% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_validity.R
\name{update_validity}
\alias{update_validity}
\title{Function to update validity variable in a \code{`observations`} table in a 
\strong{smonitor} database.}
\usage{
update_validity(con, process, summary = 0, tz = "UTC", delete = "between",
  progress = "time")
}
\arguments{
\item{con}{Database connection.}

\item{process}{A vector of processes.}

\item{summary}{Summary to invalidate. This will usually be \code{0} for source
data.}

\item{tz}{Time-zone.}

\item{delete}{Type of delete-match to use for dates.}

\item{progress}{Progress bar type. Default is \code{"time"}.}
}
\description{
\code{update_validity} does not really update the validity variable, it queries
the database, re-tests, deletes the old observations and inserts the re-tested
table.
}
\author{
Stuart K. Grange
}
