% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity_test.R
\name{validity_test}
\alias{validity_test}
\title{Function to add a validity variable to a data frame based on date ranges.}
\usage{
validity_test(df, df_look, use_process = FALSE)
}
\arguments{
\item{df}{Data frame to test and add or overwrite a \code{validity} variable.}

\item{df_look}{A look-up data frame containing \code{site}, \code{variable}, 
\code{date_start}, and \code{date_end} variables.}

\item{use_process}{Use \code{"process"} rather than \code{"site"} and 
\code{"variable"}.}
}
\description{
\code{validity_test} oes not currently handle groups, but will in the future.
It is recommended that \code{dplyr::do} is used for this task.
}
\author{
Stuart K. Grange.
}
\seealso{
\code{\link{within_range}}
}

