% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity_test.R
\name{validity_test}
\alias{validity_test}
\title{Function to add a validity variable to a data frame based on date ranges.}
\usage{
validity_test(df, df_look, valid_value = NA)
}
\arguments{
\item{df}{Data frame to test and add or overwrite a \code{validity} variable.}

\item{df_look}{A look-up data frame containing \code{process}, 
\code{date_start}, and \code{date_end} variables.}

\item{valid_value}{For values which are not invalid, \emph{i.e.} they fail 
the date range test, what value should they take? Default is \code{NA}, but
\code{1} could be appropriate too.}
}
\description{
Function to add a validity variable to a data frame based on date ranges.
}
\seealso{
\code{\link{within_range}}
}
\author{
Stuart K. Grange.
}
