% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_rose.R
\name{wind_rose}
\alias{wind_rose}
\title{Function to plot wind roses.}
\usage{
wind_rose(df, ws = "ws", wd = "wd", ws_breaks = seq(0, 10, 2),
  wd_width = 30)
}
\arguments{
\item{df}{Data frame.}

\item{ws}{Wind speed variable.}

\item{wd}{Wind direction variable}

\item{ws_breaks}{Breaks for wind speed binning.}

\item{wd_width}{Width of breaks for wind direction binning.}
}
\description{
\code{wind_rose} uses \strong{ggplot2} and is a stacked bar chart with
polar coordinates. \code{wind_rose} bins wind speed and direction then 
aggregates to find counts.
}
\author{
Stuart K. Grange
}
\seealso{
\code{\link{windRose}}
}

